comopnents for communication
